+++ noodles-csi/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * csi/index/header: Change column indices to be 0-based.
+
 ## 0.17.0 - 2023-05-04
 
 ### Added