+++ noodles-tabix/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Changed
 
+  * async/reader: Ensure header column indices are > 0.
+
   * reader: Ensure header column indices are > 0.
 
 ## 0.20.0 - 2023-05-04